[MIG] sale_order_line_description: Migration to 11.0
