+++ sale_order_line_description/README.rst
@@ -1,11 +1,37 @@
-.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
-   :target: https://www.gnu.org/licenses/agpl
-   :alt: License: AGPL-3
-
+===========================
 Sale order line description
 ===========================
 
-This module sets to use only product's sale description field on the sale order lines, instead of "Ref+Name+Sale Description".
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fsale--workflow-lightgray.png?logo=github
+    :target: https://github.com/OCA/sale-workflow/tree/11.0/sale_order_line_description
+    :alt: OCA/sale-workflow
+.. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
+    :target: https://translation.odoo-community.org/projects/sale-workflow-11-0/sale-workflow-11-0-sale_order_line_description
+    :alt: Translate me on Weblate
+.. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
+    :target: https://runbot.odoo-community.org/runbot/167/11.0
+    :alt: Try me on Runbot
+
+|badge1| |badge2| |badge3| |badge4| |badge5| 
+
+This module sets to use only product's sale description field on the sale order
+lines, instead of "Ref+Name+Sale Description".
+
+**Table of contents**
+
+.. contents::
+   :local:
 
 Configuration
 =============
@@ -13,46 +39,62 @@ Configuration
 The user has to belong to group_use_product_description_per_so_line.
 This is possible by selecting the related option in the following menu:
 
-* Sales --> Settings --> Quotations & Sales -->
-  Allow using only the product sale description on the sales order lines
+* *Sales > Settings > Quotations & Orders > Product Sale Description*
+
+Usage
+=====
+
+#. Go to **Sales > Settings > Quotations & Orders** and check "Product Sale
+   description* to use only product sale description on the sales order lines.
+#. Go to **Setting > Users & Companies > Users** and check you have checked
+   the "Use only product purchase description on order lines" access right for
+   your user.
+#. Add or modify a sale description to any of your products.
+#. Create a sale order and add a sale order line with this product and check
+   out that now sale order line description only contains the description you
+   are created.
 
 Bug Tracker
 ===========
 
-Bugs are tracked on `GitHub Issues
-<https://github.com/OCA/sale-workflow/issues>`_. In case of trouble, please
-check there if your issue has already been reported. If you spotted it first,
-help us smash it by providing detailed and welcomed feedback.
+Bugs are tracked on `GitHub Issues <https://github.com/OCA/sale-workflow/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/OCA/sale-workflow/issues/new?body=module:%20sale_order_line_description%0Aversion:%2011.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
+Do not contact contributors directly about support or help with technical issues.
 
 Credits
 =======
 
-Images
-------
+Authors
+~~~~~~~
 
-* Odoo Community Association: `Icon <https://github.com/OCA/maintainer-tools/blob/master/template/module/static/description/icon.svg>`_.
+* Agile Business Group
 
 Contributors
-------------
+~~~~~~~~~~~~
 
 * Alex Comba <alex.comba@agilebg.com>
 * Daniel Campos <danielcampos@avanzosc.es>
 * Simone Rubino <simone.rubino@agilebg.com>
+* `Tecnativa <https://www.tecnativa.com>`_ :
 
-Do not contact contributors directly about support or help with technical issues.
+  * Vicent Cubells
 
-Maintainer
-----------
-
-.. image:: http://odoo-community.org/logo.png
-   :alt: Odoo Community Association
-   :target: http://odoo-community.org
+Maintainers
+~~~~~~~~~~~
 
 This module is maintained by the OCA.
+
+.. image:: https://odoo-community.org/logo.png
+   :alt: Odoo Community Association
+   :target: https://odoo-community.org
 
 OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
 
-To contribute to this module, please visit https://odoo-community.org.
+This module is part of the `OCA/sale-workflow <https://github.com/OCA/sale-workflow/tree/11.0/sale_order_line_description>`_ project on GitHub.
+
+You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
