+++ sale_order_line_description/readme/CONTRIBUTORS.rst
@@ -1,0 +1,6 @@
+* Alex Comba <alex.comba@agilebg.com>
+* Daniel Campos <danielcampos@avanzosc.es>
+* Simone Rubino <simone.rubino@agilebg.com>
+* `Tecnativa <https://www.tecnativa.com>`_ :
+
+  * Vicent Cubells
