+++ sale_order_line_description/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+This module sets to use only product's sale description field on the sale order
+lines, instead of "Ref+Name+Sale Description".
