+++ sale_order_line_description/security/sale_security.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
 
-        <record id="group_use_product_description_per_so_line" model="res.groups">
-            <field name="name">Use only product sale description on order lines</field>
-            <field name="category_id" ref="base.module_category_hidden"/>
-        </record>
+    <record id="group_use_product_description_per_so_line" model="res.groups">
+        <field name="name">Use only product sale description on order lines</field>
+        <field name="category_id" ref="base.module_category_hidden"/>
+    </record>
 
 </odoo>
