+++ sale_order_line_description/tests/__init__.py
@@ -1,4 +1,1 @@
-# -*- coding: utf-8 -*-
-# Copyright 2017 Simone Rubino - Agile Business Group
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 from . import test_sale
