--- sale_order_line_description/views/res_config_view.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-        <record id="view_sales_config" model="ir.ui.view">
-            <field name="name">sale settings</field>
-            <field name="model">sale.config.settings</field>
-            <field name="inherit_id" ref="sale.view_sales_config"/>
-            <field name="arch" type="xml">
-                <xpath expr="//group[@name='quotations_sales']" position="inside">
-                    <label for="group_use_product_description_per_so_line" string="Product sale description"/>
-                    <div>
-                        <field name="group_use_product_description_per_so_line" class="oe_inline"/>
-                        <label for="group_use_product_description_per_so_line"/>
-                    </div>
-                </xpath>
-            </field>
-        </record>
-
-</odoo>
